Se agregan archivos al gitignore

+++ nlppen/seleccion.py
@@ -1,7 +1,7 @@
 import re
-from copy import deepcopy
 import json
 import os
+from copy import deepcopy
 
 import shutil
 
@@ -12,8 +12,9 @@ from .spark_udfs import *
 
 class Seleccion:
     sdf = None
+
     def __init__(self, terminos, spark, parquet_path='../../datasets/complete',
-                 cambios_path='./cambios.json', datasets_path='./datasets'):
+                 cambios_path='./cambios.json', datasets_path='./datasets', filtro=None):
         """ Retorna un objeto Seleccion
 
         Crea un nuevo objeto que contiene un diccioario de terminos para ser aplicados a un conjunto de datos.
@@ -27,7 +28,7 @@ class Seleccion:
 
         spark:  SparkContext
             Es el contexto de una sesión de Spark. Se encargará de leer los datos desde un parquet.
-        
+
         parquet_path: String
             Directorio donde se encuentran ubicados los archivos .parquet.
 
@@ -41,7 +42,7 @@ class Seleccion:
         self.spark = spark
         self.parquet_path = parquet_path
         self.datasets_path = datasets_path
-
+        self.filtro_carga = filtro
 
         if os.path.exists(cambios_path):
             with open(cambios_path) as f:
@@ -52,43 +53,58 @@ class Seleccion:
         if not os.path.exists(datasets_path):
             os.makedirs(datasets_path)
 
-        self.sdf = self.spark.read.parquet(self.parquet_path)
+        if self.filtro_carga is None:
+            self.sdf = self.spark.read.parquet(self.parquet_path)
+        else:
+            self.sdf = self.spark.read.parquet(
+                self.parquet_path).where(self.filtro_carga)
 
     def cargar_datos(self):
         if os.path.exists(self.parquet_path):
-            self.sdf = self.spark.read.parquet(self.parquet_path)
-    
-    def guardarDatos(self, parquet_file='terminos.parquet'):
+            if self.filtro_carga is None:
+                self.sdf = self.spark.read.parquet(self.parquet_path)
+            else:
+                self.sdf = self.spark.read.parquet(
+                    self.parquet_path).where(self.filtro_carga)
+
+    def guardarDatos(self, parquet_file='terminos.parquet', partition_cols=None, borrar=True):
         if self.sdf is not None:
             parquet_path = self.datasets_path + '/' + parquet_file
-            if os.path.exists(parquet_path):
-                 shutil.rmtree(parquet_path)
-            self.sdf.write.parquet(parquet_path)
-            
-    def cargarPreprocesados(self, parquet_file='terminos.parquet'):
+            if borrar and os.path.exists(parquet_path):
+                shutil.rmtree(parquet_path)
+
+            if partition_cols is None:
+                self.sdf.write.parquet(parquet_path)
+            else:
+                (self.sdf.repartition(*partition_cols)
+                .write.mode('append').partitionBy(*partition_cols).parquet(parquet_path))
+
+    def cargarPreprocesados(self, parquet_file='terminos.parquet', filtro=None):
         parquet_path = self.datasets_path + '/' + parquet_file
         if os.path.exists(parquet_path):
-            self.sdf = self.spark.read.parquet(parquet_path)
+            if filtro is None:
+                self.sdf = self.spark.read.parquet(parquet_path)
+            else:
+                self.sdf = self.spark.read.parquet(parquet_path).where(filtro)
 
-    
-    def filtrar_sentencias(self, parquet_file='terminos.parquet', preprocess=None, save=True, keepRowEmpty=False):
+    def filtrar_sentencias(self, parquet_file='terminos.parquet', precargar=True, preprocess=None, save=True, keepRowEmpty=False, partition_cols=None):
         """ Sobreescribe el Dataframe aplicando el filtro de las sentencias de acuerdo a los terminos de la clase.
 
         Parametros:
 
         parquet_file: String
             Nombre del directorio donde se almacenaras los archivos parquet correspondientes a las sentencias filtradas, si ya existe los lee.
-        
+
         preprocess : function
             Una funcion para aplicar como preprocesamiento del dataset, antes de aplicar los terminos (e.g solo_portanto())
         """
         parquet_path = self.datasets_path + '/' + parquet_file
-        if os.path.exists(parquet_path):
+        if precargar and os.path.exists(parquet_path):
             self.sdf = self.spark.read.parquet(parquet_path)
         else:
             self.__busqueda_terminos(preprocess, keepRowEmpty)
             if save:
-                self.guardarDatos(parquet_file)
+                self.guardarDatos(parquet_file, partition_cols=partition_cols)
 
         return self.sdf
 
@@ -113,18 +129,17 @@ class Seleccion:
         df.columns = [term.replace('_', ' ') for term in df.columns]
         df.index = [term.replace('_', ' ') for term in df.index]
         return df
-    
 
     def tabla_terminos_anno(self):
         terminos_cols = [col.replace(' ', '_') for col in self.terminos.keys()]
         ldf = (self.sdf
-        .select('anno', *terminos_cols)
-        .groupby('anno').sum()
-        .sort('anno')
-        .toPandas()
-        .set_index('anno')
-        .drop(columns=['sum(anno)'])
-        )
+               .select('anno', *terminos_cols)
+               .groupby('anno').sum()
+               .sort('anno')
+               .toPandas()
+               .set_index('anno')
+               .drop(columns=['sum(anno)'])
+               )
         ldf.columns = self.terminos.keys()
         return ldf
 
@@ -148,19 +163,22 @@ class Seleccion:
         """
         schema = deepcopy(self.sdf.schema)
 
-        term_regex = {} # Nuevo diccionario para contener los terminos y sus expresiones regulares compiladas.
+        # Nuevo diccionario para contener los terminos y sus expresiones regulares compiladas.
+        term_regex = {}
         for cat, lst in self.terminos.items():
             col_name = cat.replace(' ', '_')
             term_regex[col_name] = [re.compile(r'\s+' + t.replace(' ', r'[\s\.\,\-\)\;\:\]]+'),  re.X | re.M | re.I)
-                                    for t in lst] # Compila las experesiones regulares
-            schema.add(col_name, 'integer', True) # Agregar las nuevas columnas
+                                    for t in lst]  # Compila las experesiones regulares
+            # Agregar las nuevas columnas
+            schema.add(col_name, 'integer', True)
 
         self.sdf = (self.sdf.rdd
-                    .map(lambda row: spark_buscar_terminos_doc(row, term_regex, preprocess=preprocess, keepRowEmpty=keepRowEmpty)) # Ejecuta la busqueda de terminos con spark
+                    # Ejecuta la busqueda de terminos con spark
+                    .map(lambda row: spark_buscar_terminos_doc(row, term_regex, preprocess=preprocess, keepRowEmpty=keepRowEmpty))
                     .filter(lambda d: d is not None)
-                    .toDF(schema=schema) 
+                    .toDF(schema=schema)
                     .persist()
-                    ) # Guarda el nuevo esquema
+                    )  # Guarda el nuevo esquema
         return self.sdf
 
     def sub_busqueda(self, terminos_sub, actualizar_sdf=False, preprocess=None, keepRowEmpty=False):
@@ -178,7 +196,7 @@ class Seleccion:
                     Es un diccionario de terminos a ser buscados en el conjunto de sentencias. 
                     Cada llave será agregada como una columna del conjunto de datos, mientras que cada valor
                     para el cojunto de valores representa una expresión regular.
-                
+
                 actualizar_sdf: Booleano.
                     True para reescribir el sdf.
                     False para no sobreescribir el sdf.
@@ -187,21 +205,21 @@ class Seleccion:
         schema = deepcopy(self.sdf.schema)
         term_regex = {}
 
-        #Crear las nuevas columnas
+        # Crear las nuevas columnas
         for cat, lst in terminos_sub.items():
             col_name = cat.replace(' ', '_')
             term_regex[col_name] = [re.compile(r'\s+' + t.replace(' ', r'[\s\.\,\-\)\;\:\]]+'))
                                     for t in lst]
             schema.add(col_name, 'integer', True)
-        
-        #Aplicar la busqueda de terminos
+
+        # Aplicar la busqueda de terminos
         self.subbusqueda = (self.sdf.rdd
                             .map(lambda row: spark_buscar_terminos_doc(row, term_regex, preprocess=preprocess, keepRowEmpty=keepRowEmpty))
                             .filter(lambda d: d is not None)
                             .toDF(schema=schema)
                             .persist()
                             )
-        #Verificar para actualizar el sdf
+        # Verificar para actualizar el sdf
         if actualizar_sdf:
             self.terminos = {**self.terminos, **terminos_sub}
             self.sdf = self.subbusqueda
